imprv: divide data table component to provider, root, body, header to enable customization
