+++ src/components/ui/data-table/data-table-provider.tsx
@@ -1,0 +1,70 @@
+'use client';
+
+import {
+  ColumnDef,
+  ColumnFiltersState,
+  getCoreRowModel,
+  getFilteredRowModel,
+  getPaginationRowModel,
+  getSortedRowModel,
+  PaginationState,
+  SortingState,
+  Table as TableType,
+  useReactTable,
+  VisibilityState,
+} from '@tanstack/react-table';
+import { useState } from 'react';
+
+interface DataTableProps<TData, TValue> {
+  columns: ColumnDef<TData, TValue>[];
+  data: TData[];
+  children: (props: {
+    table: TableType<TData>;
+    columns: ColumnDef<TData, TValue>[];
+  }) => React.ReactNode;
+}
+
+// @TODO: add enablePagination prop and make it context value, and use it inside data-table-pagination component
+export const DataTableProvider = <TData, TValue>({
+  columns,
+  data,
+  children,
+}: DataTableProps<TData, TValue>) => {
+  const [sorting, setSorting] = useState<SortingState>([]);
+  const [columnFilters, setColumnFilters] = useState<ColumnFiltersState>([]);
+  const [columnVisibility, setColumnVisibility] = useState<VisibilityState>({});
+  const [rowSelection, setRowSelection] = useState({});
+  const [pagination, setPagination] = useState<PaginationState>({
+    pageIndex: 0,
+    pageSize: 10,
+  });
+
+  // @TODO: make it compounded component that spread table object
+  const table = useReactTable({
+    data,
+    columns,
+    onSortingChange: setSorting,
+    onColumnFiltersChange: setColumnFilters,
+    getCoreRowModel: getCoreRowModel(),
+    getPaginationRowModel: getPaginationRowModel(),
+    getSortedRowModel: getSortedRowModel(),
+    getFilteredRowModel: getFilteredRowModel(),
+    onColumnVisibilityChange: setColumnVisibility,
+    onRowSelectionChange: setRowSelection,
+    onPaginationChange: setPagination,
+    state: {
+      sorting,
+      columnFilters,
+      columnVisibility,
+      rowSelection,
+      pagination,
+    },
+  });
+
+  return (
+    <>
+      {/* pass more state if needed */}
+      {children({ table, columns })}
+    </>
+  );
+};
