+++ src/components/ui/data-table/data-table-root.tsx
@@ -1,0 +1,9 @@
+import { ComponentPropsWithoutRef, FC } from 'react';
+
+import { Table } from '@/components/ui/table';
+
+type DataTableRootProps = ComponentPropsWithoutRef<typeof Table>;
+
+export const DataTableRoot: FC<DataTableRootProps> = (props) => {
+  return <Table {...props} />;
+};
